ref: moved commands into speperate file
