+++ cli/src/command/mod.rs
@@ -1,0 +1,5 @@
+pub mod definition;
+pub mod download;
+pub mod feature;
+pub mod report;
+pub mod watch;
